refactor(classic): extract the shared classic page shell
The catalog and library shells were byte-identical trees — Navigation over centered full-width content, same stylesheet import, same comment about tubafrenzy's missing .centerWidth cap. Both now consume one shell at src/components/experiences/classic/Layout/Main.tsx.
