+++ app/dashboard/@classic/catalog/page.tsx
@@ -1,6 +1,6 @@
 import { Metadata } from "next";
 import { getPageTitle } from "@/lib/utils/page-title";
-import Main from "@/src/components/experiences/classic/catalog/Layout/Main";
+import Main from "@/src/components/experiences/classic/Layout/Main";
 import SearchForm from "@/src/components/experiences/classic/catalog/SearchForm";
 import SearchResults from "@/src/components/experiences/classic/catalog/SearchResults";
